3143.   [bug]           Silence clang compiler warnings. [RT #25174]

+++ lib/dns/dst_api.c
@@ -31,7 +31,7 @@
 
 /*
  * Principal Author: Brian Wellington
- * $Id: dst_api.c,v 1.16.12.13 2011/08/18 05:14:33 marka Exp $
+ * $Id: dst_api.c,v 1.16.12.14 2011/08/18 06:40:45 marka Exp $
  */
 
 /*! \file */
@@ -1193,7 +1193,7 @@ write_public_key(const dst_key_t *key, int type, const char *directory) {
 	fprintf(fp, " ");
 
 	isc_buffer_usedregion(&classb, &r);
-       if (isc_util_fwrite(r.base, 1, r.length, fp) != r.length)
+	if (isc_util_fwrite(r.base, 1, r.length, fp) != (int)r.length)
                ret = DST_R_WRITEERROR;
 
 	if ((type & DST_TYPE_KEY) != 0)
@@ -1202,7 +1202,7 @@ write_public_key(const dst_key_t *key, int type, const char *directory) {
 		fprintf(fp, " DNSKEY ");
 
 	isc_buffer_usedregion(&textb, &r);
-       if (isc_util_fwrite(r.base, 1, r.length, fp) != r.length)
+	if (isc_util_fwrite(r.base, 1, r.length, fp) != (int)r.length)
                ret = DST_R_WRITEERROR;
 
 	fputc('\n', fp);
